requirements.txt
flask>=3.0
python-dotenv>=1.0
requests>=2.31
pytest>=8.0
pytest-cov>=5.0
